feat: added support for both mdx and raw csf
This 'adds back' some stuff to explore if I can still use mdx2csf.
The main issue I have with the other approach (the non-mdx approach) is the styling overhead I need to do to get it to look decent in the sidebar, and the fact that it will be my component (rather than pre-built storybook one).