refactor: remove mdx indexing + tidy up csf/compile
This is the removal of the mdx way of doing things (i.e. config -> mdx -> csf) in favour of just (config -> csf). I think its safe to say doing it via stories directly facilitates more flexibility - but this can be reverted if so.

I'm also including the addition of the Typopgraphy base story to the csf file here ; with functionality to be added later (and removing whats already there for now).