feat: added version detection based on file path
This adds the ability to use the v3 config when its a tailwind.config.j|ts file, and v4 config when its .css file; so no need to pass in the version manually.

I also moved the helpers to a folder in an effort to start structuring things properly as we move from the prototyping to refining stage.