test: added test to show resolveConfig behaviour
This test isn't testing a big chunk of logic, and is instead intended to sort of show what resolveConfig is meant to do. It also represents what the structure should look like at the end. I'm still not sure on it - so leaving a TODO and committing this for now.