refactor: change resolveId to base one + remove webpack
In our load function in unplugin.js, we remove the js extensions. To make it clear as to where this is extention is first added, I've updated the resolveId method to not include the js extension at the Loader level, and instead at it within unplugin, and updated tests accordingly.
I've also decided I don't want to support Webpack yet, and will leave it for the roadmap.